+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - CLI implemented using `entry_points` instead of a dedicated script.
 
+### Fixed
+- Entry type inference now correct by inheriting from `object`.
+
 ## [0.1.0] - 2016-12-16
 ### Added
 - Basic data structures